requirements.txt
flask
google-generativeai
moviepy
openai-whisper
Werkzeug
python-dotenv
google.generativeai